# examples/OpenAPI_examples/SmartDoor_actuator/SmartDoor_OpenAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Web Thing Model",
    "description": "This is a Web Thing Model server. You can find\nout more about Web Thing Model (W3C) at\n[https://www.w3.org/Submission/wot-model/](https://www.w3.org/Submission/wot-model/).\n",
    "contact": {
      "email": "atzavaras@isc.tuc.gr"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "http://172.16.1.1:5002/SmartDoor_actuator",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
...
  "paths": {
...
    "/properties": {
      "get": {
        "tags": [
          "Properties"
        ],
        "summary": "Retrieve a list of properties",
        "description": "In response to an HTTP GET request on the destination URL of a properties link, an Extended Web Thing must return an array of Property that the initial resource contains.",
        "operationId": "retrieveWebThingProperties",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PropertiesResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/actions": {
      "get": {
        "tags": [
          "Actions"
        ],
        "summary": "Retrieve a list of actions",
        "operationId": "retrieveWebThingActions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Action"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PropertiesResponse": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/StateProperty"
          }
        ],
        "xml": {
          "name": "PropertiesResponse"
        }
      },
      "State": {
        "type": "object",
        "x-kindOf": "http://www.w3.org/ns/sosa/Observation",
        "properties": {
          "state": {
            "type": "string",
            "x-kindOf": "http://www.w3.org/ns/sosa/hasSimpleResult",
            "enum": [
              "OPEN",
              "CLOSED",
              "LOCKED"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "x-kindOf": "http://www.w3.org/ns/sosa/resultTime"
          }
        },
        "xml": {
          "name": "State"
        }
      },
      "StateProperty": {
        "required": [
          "id",
          "values"
        ],
        "x-kindOf": "http://www.w3.org/ns/ssn/systems/SystemCapability",
        "properties": {
          "id": {
            "type": "string",
            "default": "state",
            "x-kindOf": "http://schema.org/identifier"
          },
          "name": {
            "type": "string",
            "example": "Smart door's current state",
            "x-kindOf": "http://schema.org/name"
          },
          "values": {
            "$ref": "#/components/schemas/State"
          }
        },
        "xml": {
          "name": "StateProperty"
        }
      },
      "Action": {
        "required": [
          "id",
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "lock"
          },
          "name": {
            "type": "string",
            "example": "Lock the Smart Door"
          }
        },
        "xml": {
          "name": "Action"
        }
      },
...
    },
...
  }
}